Proctoring-AI-BE-M4/Proctoring-AI-BE-M4/requirements.txt
# Note: CMake must be installed first:
# macOS: brew install cmake
# Ubuntu: apt-get install cmake
# Windows: Download from cmake.org

# Python version 3.9

# Core dependencies
flask==2.2.5
requests==2.31.0
numpy==1.24.3
pandas>=2.0.3
scikit-learn>=1.0.2
matplotlib>=3.5.3

# FastAPI and related
fastapi==0.103.2
uvicorn==0.23.2
gunicorn==20.1.0
websockets==11.0.3
pydantic>=2.4.2
pydantic-settings>=2.0.3

# ML/CV dependencies
ultralytics>=8.0.0,<9.0.0
opencv-python-headless==4.10.0.84  # Updated to match Roboflow requirement
mediapipe==0.10.5
torch>=2.0.1
torchvision>=0.15.2

# Face Recognition
tensorflow-cpu==2.15.0
tf-keras==2.15.0
keras>=2.15.0
# deepface and retina-face installed separately in Dockerfile with --no-deps
gdown>=3.10.1
mtcnn>=0.1.0
pillow>=10.0.0
tqdm>=4.65.0
openpyxl>=3.1.2

# Database and auth
redis>=4.5.0,<6.0.0
mysql-connector-python==8.0.33
psycopg2-binary==2.9.9  # Added for PostgreSQL support
SQLAlchemy==2.0.23
python-multipart==0.0.6
PyLTI1p3>=1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.0.0
roboflow>=1.1.58
boto3>=1.28.0

# Configuration
python-dotenv==1.0.0

# Observability
prometheus-client>=0.20.0
